feat(loans): approve a loan request given a valid amount
- update [server/db/migrations/20190517000400-create-staff.js, server/models/staff.js] add secreteKey collumn for JWT auth and revoke purposes in the future
update necessary database columns to use decimal for everything relating to money [server/db/migrations/20190517180907-create-account-type.js, server/db/migrations/20190517224621-create-loan-type.js, server/db/migrations/20190528125318-create-loan.js, server/models/loan.js, server/models/loantype.js]
- update server/db/seeders/seedSuperAdmin.js to reflect the new secrectKey column
- update server/helpers/generateAuthToken.js to reflect the secrectKey
- update server/middlewares/checkAccessLevel.js: add isAdmin accessLevel
- update server/middlewares/checkAuth.js: check the validity of the secretKey and check the token expiry
- update server/middlewares/validation/loanMethods.js: add validateApproveLoan method, update validateCreateLoan to reflect the duration
- update server/middlewares/validation/validateReqKeys.js: add some more params to validate
- update server/middlewares/validationMiddleware.js: validate that minimumAmount is > than maximumAmount in validateCreateLoanType method

- update [server/models/loan.js, server/db/migrations/20190528125318-create-loan.js] add duration and approvalStatus columns

[Finishes #166325647]

+++ server/models/accounttype.js
@@ -4,7 +4,7 @@ export default (sequelize, DataTypes) => {
     {
       name: DataTypes.STRING,
       interestRate: DataTypes.INTEGER,
-      minimumBalance: DataTypes.INTEGER,
+      minimumBalance: DataTypes.DECIMAL,
     },
     {},
   );
